+++ internal/config/config.go
@@ -5,6 +5,6 @@ const (
 	GithubLink = "https://github.com/avm-collection/anasm"
 
 	VersionMajor = 1
-	VersionMinor = 15
+	VersionMinor = 17
 	VersionPatch = 9
 )